Update base image of nop Dockerfile
Sidecar tests are failing on CI.
The tests fails when the nop images are involved.
The nop image built by CI uses `ubi` as the base image as it specified so
in the generated Dockerfile for nop image

In local testing I could observe that the sidecar tests pass whe `scratch` is used
as the base image for nop. Upstream also seems to be using scratch as their base image for `nop` image

Hence, the purpiose of this patch to test whether changing the nop image base to `scratch` will make the sidecar test pass on ci